# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
      "AccentPhrase": {
        "description": "アクセント句ごとの情報",
        "properties": {
          "accent": {
            "title": "アクセント箇所",
            "type": "integer"
          },
          "is_interrogative": {
            "default": false,
            "title": "疑問系かどうか",
            "type": "boolean"
          },
          "moras": {
            "items": {
              "$ref": "#/components/schemas/Mora"
            },
            "title": "モーラのリスト",
            "type": "array"
          },
          "pause_mora": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Mora"
              },
              {
                "type": "null"
              }
            ],
            "title": "後ろに無音を付けるかどうか"
          }
        },
        "required": [
          "moras",
          "accent"
        ],
        "title": "AccentPhrase",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParseKanaBadRequest": {
        "properties": {
          "error_args": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "エラーを起こした箇所",
            "type": "object"
          },
          "error_name": {
            "description": "|name|description|\n|---|---|\n| UNKNOWN_TEXT | 判別できない読み仮名があります: {text} |\n| ACCENT_TOP | 句頭にアクセントは置けません: {text} |\n| ACCENT_TWICE | 1つのアクセント句に二つ以上のアクセントは置けません: {text} |\n| ACCENT_NOTFOUND | アクセントを指定していないアクセント句があります: {text} |\n| EMPTY_PHRASE | {position}番目のアクセント句が空白です |\n| INTERROGATION_MARK_NOT_AT_END | アクセント句末以外に「？」は置けません: {text} |\n| INFINITE_LOOP | 処理時に無限ループになってしまいました...バグ報告をお願いします。 |",
            "title": "エラー名",
            "type": "string"
          },
          "text": {
            "title": "エラーメッセージ",
            "type": "string"
          }
        },
        "required": [
          "text",
          "error_name",
          "error_args"
        ],
        "title": "ParseKanaBadRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/accent_phrases": {
      "post": {
        "description": "テキストからアクセント句を得ます。\nis_kanaが`true`のとき、テキストは次の AquesTalk 風記法で解釈されます。デフォルトは`false`です。\n* 全てのカナはカタカナで記述される\n* アクセント句は`/`または`、`で区切る。`、`で区切った場合に限り無音区間が挿入される。\n* カナの手前に`_`を入れるとそのカナは無声化される\n* アクセント位置を`'`で指定する。全てのアクセント句にはアクセント位置を1つ指定する必要がある。\n* アクセント句末に`？`(全角)を入れることにより疑問文の発音ができる。",
        "operationId": "accent_phrases_accent_phrases_post",
        "parameters": [
          {
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "title": "Text",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "is_kana",
            "required": false,
            "schema": {
              "default": false,
              "title": "Is Kana",
              "type": "boolean"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccentPhrase"
                  },
                  "title": "Response Accent Phrases Accent Phrases Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseKanaBadRequest"
                }
              }
            },
            "description": "読み仮名のパースに失敗"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "テキストからアクセント句を得る",
        "tags": [
          "クエリ編集"
        ]
      }
    },
...
    "/audio_query": {
      "post": {
        "description": "音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。",
        "operationId": "audio_query_audio_query_post",
        "parameters": [
          {
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "title": "Text",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioQuery"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "音声合成用のクエリを作成する",
        "tags": [
          "クエリ作成"
        ]
      }
    },
    "/audio_query_from_preset": {
      "post": {
        "description": "音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。",
        "operationId": "audio_query_from_preset_audio_query_from_preset_post",
        "parameters": [
          {
            "in": "query",
            "name": "text",
            "required": true,
            "schema": {
              "title": "Text",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioQuery"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "音声合成用のクエリをプリセットを用いて作成する",
        "tags": [
          "クエリ作成"
        ]
      }
    },
    "/cancellable_synthesis": {
      "post": {
        "operationId": "cancellable_synthesis_cancellable_synthesis_post",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "音声合成"
        ]
      }
    },
    "/connect_waves": {
      "post": {
        "description": "base64エンコードされたwavデータを一纏めにし、wavファイルで返します。",
        "operationId": "connect_waves_connect_waves_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "title": "Waves",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "base64エンコードされた複数のwavデータを一つに結合する",
        "tags": [
          "その他"
        ]
      }
    },
...
    "/frame_synthesis": {
      "post": {
        "operationId": "frame_synthesis_frame_synthesis_post",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrameAudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "音声合成"
        ]
      }
    },
...
    "/mora_data": {
      "post": {
        "operationId": "mora_data_mora_data_post",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AccentPhrase"
                },
                "title": "Accent Phrases",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccentPhrase"
                  },
                  "title": "Response Mora Data Mora Data Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "アクセント句から音高・音素長を得る",
        "tags": [
          "クエリ編集"
        ]
      }
    },
    "/mora_length": {
      "post": {
        "operationId": "mora_length_mora_length_post",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AccentPhrase"
                },
                "title": "Accent Phrases",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccentPhrase"
                  },
                  "title": "Response Mora Length Mora Length Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "アクセント句から音素長を得る",
        "tags": [
          "クエリ編集"
        ]
      }
    },
...
  }
}